+++ Mu3Library_Base/Runtime/Scripts/DI/CoreRoot.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Mu3Library.Foundation.Event;
 using UnityEngine;
 
 namespace Mu3Library.DI
@@ -57,7 +58,10 @@ namespace Mu3Library.DI
 
         private readonly Dictionary<Type, CoreBase> _cores = new();
         private readonly List<CoreBase> _orderedCores = new();
-        internal event Action<Type> OnCoreAdded;
+
+        private readonly SubscribeHandler _subscribeHandler = new();
+        public event Action<Type> OnCoreInitialized;
+        public event Action<Type> OnCorePrepared;
 
 
 
@@ -128,9 +132,11 @@ namespace Mu3Library.DI
 
             _orderedCores.Add(core);
             _orderedCores.Sort(CompareCoreOrder);
+
+            core.SubscribeOnInitializedOnce(() => OnCoreInitialized?.Invoke(type));
             core.InitializeCore();
 
-            OnCoreAdded?.Invoke(type);
+            core.SubscribeOnPreparedOnce(() => OnCorePrepared?.Invoke(type));
         }
         #endregion
 
@@ -169,6 +175,88 @@ namespace Mu3Library.DI
         {
             Type type = typeof(T);
             return _cores.ContainsKey(type);
+        }
+
+        public ISubscriptionInfo SubscribeOnCoreInitializedOnce<T>(Action callback) where T : CoreBase
+        {
+            Type type = typeof(T);
+            return SubscribeOnCoreInitializedOnce(type, callback);
+        }
+
+        public ISubscriptionInfo SubscribeOnCoreInitializedOnce(Type type, Action callback)
+        {
+            if (type == null)
+            {
+                return null;
+            }
+
+            if (_cores.TryGetValue(type, out var core))
+            {
+                return core.SubscribeOnInitializedOnce(callback);
+            }
+            else
+            {
+                ISubscriptionInfo subscription = null;
+                Action<Type> handler = initializedType =>
+                {
+                    if (initializedType != type)
+                    {
+                        return;
+                    }
+
+                    _subscribeHandler.Deregister(subscription);
+                    callback?.Invoke();
+                };
+
+                subscription = _subscribeHandler.Register(
+                    () => OnCoreInitialized += handler,
+                    () => OnCoreInitialized -= handler
+                );
+                subscription?.Subscribe();
+
+                return subscription;
+            }
+        }
+
+        public ISubscriptionInfo SubscribeOnCorePreparedOnce<T>(Action callback) where T : CoreBase
+        {
+            Type type = typeof(T);
+            return SubscribeOnCorePreparedOnce(type, callback);
+        }
+
+        public ISubscriptionInfo SubscribeOnCorePreparedOnce(Type type, Action callback)
+        {
+            if (type == null)
+            {
+                return null;
+            }
+
+            if (_cores.TryGetValue(type, out var core))
+            {
+                return core.SubscribeOnPreparedOnce(callback);
+            }
+            else
+            {
+                ISubscriptionInfo subscription = null;
+                Action<Type> handler = initializedType =>
+                {
+                    if (initializedType != type)
+                    {
+                        return;
+                    }
+
+                    _subscribeHandler.Deregister(subscription);
+                    callback?.Invoke();
+                };
+
+                subscription = _subscribeHandler.Register(
+                    () => OnCorePrepared += handler,
+                    () => OnCorePrepared -= handler
+                );
+                subscription?.Subscribe();
+
+                return subscription;
+            }
         }
         #endregion