+++ Mu3Library_Base/Runtime/Scripts/DI/ICoreRoot.cs
@@ -1,4 +1,5 @@
 using System;
+using Mu3Library.Foundation.Event;
 
 namespace Mu3Library.DI
 {
@@ -11,5 +12,16 @@ namespace Mu3Library.DI
         public object GetClass(Type coreType, Type serviceType, string key);
 
         public bool HasCore<T>() where T : CoreBase;
+
+        public event Action<Type> OnCoreInitialized;
+        public event Action<Type> OnCorePrepared;
+
+
+
+        public ISubscriptionInfo SubscribeOnCoreInitializedOnce<T>(Action callback) where T : CoreBase;
+        public ISubscriptionInfo SubscribeOnCoreInitializedOnce(Type type, Action callback);
+
+        public ISubscriptionInfo SubscribeOnCorePreparedOnce<T>(Action callback) where T : CoreBase;
+        public ISubscriptionInfo SubscribeOnCorePreparedOnce(Type type, Action callback);
     }
 }