+++ Mu3Library_Base/Runtime/Scripts/DI/IDICore.cs
@@ -3,6 +3,8 @@ namespace Mu3Library.DI
     public interface IDICore
     {
         public bool IsContainerConfigured { get; }
+        public bool IsInitialized { get; }
+        public bool IsPreparing { get; }
         public bool IsPrepared { get; }
 
         public int ExecutionOrder { get; }